hypixel scoreboard tablist data and skyblock utils working
