+++ src/main/java/at/hannibal2/skyhanni/mixins/transformers/AccessorGuiPlayerTabOverlay.java
@@ -1,11 +1,14 @@
 package at.hannibal2.skyhanni.mixins.transformers;
 
-import net.minecraft.client.gui.GuiPlayerTabOverlay;
 import net.minecraft.util.IChatComponent;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-@Mixin(GuiPlayerTabOverlay.class)
+//#if MC < 1.21
+@Mixin(net.minecraft.client.gui.GuiPlayerTabOverlay.class)
+//#else
+//$$ @Mixin(net.minecraft.client.gui.hud.PlayerListHud.class)
+//#endif
 public interface AccessorGuiPlayerTabOverlay {
     @Accessor("footer")
     IChatComponent getFooter_skyhanni();
