+++ src/main/java/at/hannibal2/skyhanni/utils/LorenzLogger.kt
@@ -1,9 +1,7 @@
 package at.hannibal2.skyhanni.utils
 
 import at.hannibal2.skyhanni.SkyHanniMod
-//#if MC < 1.21
-import at.hannibal2.skyhanni.utils.LorenzUtils.formatCurrentTime
-//#endif
+import at.hannibal2.skyhanni.utils.TimeUtils.formatCurrentTime
 import java.io.File
 import java.io.IOException
 import java.text.SimpleDateFormat
@@ -24,10 +22,6 @@ class LorenzLogger(filePath: String) {
         private val LOG_DIRECTORY = File("config/skyhanni/logs")
         private var PREFIX_PATH: String
         var hasDone = false
-        // todo remove once lorenz utils works
-        //#if MC > 1.21
-        //$$ fun SimpleDateFormat.formatCurrentTime(): String = this.format(System.currentTimeMillis())
-        //#endif
 
         init {
             val format = SimpleDateFormat("yyyy_MM_dd/HH_mm_ss").formatCurrentTime()
