+++ src/main/java/at/hannibal2/skyhanni/utils/SimpleTimeMark.kt
@@ -10,7 +10,6 @@ import kotlin.math.abs
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 
-// todo 1.21 impl needed
 @JvmInline
 value class SimpleTimeMark(private val millis: Long) : Comparable<SimpleTimeMark> {
 
@@ -63,9 +62,7 @@ value class SimpleTimeMark(private val millis: Long) : Comparable<SimpleTimeMark
 
     fun toMillis() = millis
 
-    //#if MC < 1.21
     fun toSkyBlockTime(): SkyBlockTime = SkyBlockTime.fromTimeMark(this)
-    //#endif
 
     fun toLocalDate(): LocalDate = toLocalDateTime().toLocalDate()
 
@@ -82,10 +79,8 @@ value class SimpleTimeMark(private val millis: Long) : Comparable<SimpleTimeMark
 
         fun Long.asTimeMark() = SimpleTimeMark(this)
 
-        //#if MC < 1.21
         @Deprecated("Use toTimeMark() instead", ReplaceWith("this.toTimeMark()"))
         fun SkyBlockTime.asTimeMark() = SimpleTimeMark(toMillis())
-        //#endif
 
     }
 }
