+++ src/main/java/at/hannibal2/skyhanni/utils/SkyBlockUtils.kt
@@ -3,12 +3,14 @@ package at.hannibal2.skyhanni.utils
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.data.HypixelData
 import at.hannibal2.skyhanni.data.IslandType
+//#if MC < 1.21
 import at.hannibal2.skyhanni.features.misc.IslandAreas
 import at.hannibal2.skyhanni.test.SkyBlockIslandTest
+//#endif
 import at.hannibal2.skyhanni.test.TestBingo
-import at.hannibal2.skyhanni.utils.LorenzUtils.isInIsland
 import at.hannibal2.skyhanni.utils.compat.MinecraftCompat
 
+// todo 1.21 impl needed
 object SkyBlockUtils {
 
     val onHypixel get() = HypixelData.connectedToHypixel && MinecraftCompat.localPlayerExists
@@ -19,6 +21,8 @@ object SkyBlockUtils {
 
     val inHypixelLobby get() = onHypixel && HypixelData.inLobby
 
+
+    //#if MC < 1.21
     /**
      * Consider using [IslandType.isInIsland] instead
      */
@@ -26,6 +30,9 @@ object SkyBlockUtils {
 
     // almost always prefer this over scoreboardArea
     val graphArea get() = if (inSkyBlock) IslandAreas.currentArea else null
+    //#else
+    //$$ val currentIsland get() = HypixelData.skyBlockIsland
+    //#endif
 
     // Only use scoreboardArea if graph data is not useable in this scenario.
     val scoreboardArea get() = if (inSkyBlock) HypixelData.skyBlockArea else null
